Create Virtual Documentation
Virtual Documentation is very useful and interesting for the Government in innovative teaching method which replaces the usage of Paper Documentation that helps to interact with machines . It is very challenging in the field of  pattern recognition and image processing .Object Tracking is considered as an important work which supports to interact easily with the audience Virtually . Generally, video analysis procedure has three major steps: firstly, detecting of the object, secondly tracking its movement from frame to frame and lastly analysing the behaviour of that object. For object tracking, four different issues are taken into account; selection of suitable object representation, feature selection for tracking, object detection and object tracking. In mouse or touch pad can draw legible using mouse or touch pad .This will also help the deaf people to recognize the pattern or drawings and good for interaction in Virtual Mode